requirement.txt
Flask==3.0.3
tensorflow==2.16.1
Pillow==10.3.0
numpy==1.26.4
